+++ packages/number-input/src/Select/Select.tsx
@@ -1,22 +1,14 @@
-import React, { useState } from 'react';
+import React from 'react';
 
-import Button, { ButtonProps } from '@leafygreen-ui/button';
+import { ButtonProps } from '@leafygreen-ui/button';
 import { cx } from '@leafygreen-ui/emotion';
-import { useForwardedRef } from '@leafygreen-ui/hooks';
 import { useDarkMode } from '@leafygreen-ui/leafygreen-provider';
-import {
-  menuButtonTextClassName,
-  Option,
-  popoverClassName,
-  Select as SelectComponent,
-} from '@leafygreen-ui/select';
-import Tooltip from '@leafygreen-ui/tooltip';
+import { Option, Select as SelectComponent } from '@leafygreen-ui/select';
+
+import { UnitOption } from '../NumberInput/NumberInput.types';
+import { SelectButton } from '../SelectButton';
 
 import {
-  customMenuButtonWrapperStyles,
-  menuBaseStyles,
-  menuThemeDisabledStyles,
-  menuThemeStyles,
   selectDisabledStyles,
   selectStyles,
   wrapperBaseStyles,
@@ -28,7 +20,7 @@ import { SelectProps } from './Select.types';
  */
 export function Select({
   'data-testid': dataTestId,
-  unit: unitProp = '',
+  unit = '',
   id,
   unitOptions,
   onChange,
@@ -53,12 +45,12 @@ export function Select({
   /**
    * Gets the current unit option using the unit string
    */
-  const currentUnitOption = unitOptions?.find(
-    unit => unit.displayName === unitProp,
-  );
+  const currentUnitOption = unitOptions.find(
+    u => u.displayName === unit,
+  ) as UnitOption;
 
   const handleChange = (val: string) => {
-    const selectedUnit = unitOptions.find(unit => unit.displayName === val);
+    const selectedUnit = unitOptions.find(u => u.displayName === val);
 
     if (selectedUnit !== undefined) {
       onChange(selectedUnit);
@@ -70,72 +62,17 @@ export function Select({
    * Tooltip will show up if there is an ellipse.
    */
   const CustomMenuButton = React.forwardRef(
-    ({ className, children, ...props }: ButtonProps, forwardedRef) => {
-      const [open, setOpen] = useState<boolean>(false);
-      const buttonRef: React.MutableRefObject<HTMLElement | null> =
-        useForwardedRef(
-          forwardedRef,
-          null,
-        ) as React.MutableRefObject<HTMLElement | null>;
-
-      /**
-       * Gets the text node for the selected option.
-       */
-      const textNode = buttonRef.current?.querySelector(
-        `.${menuButtonTextClassName}`,
-      ) as HTMLElement;
-
-      /**
-       * Checks if the selected option has an ellipse.
-       */
-      const hasEllipse = textNode?.offsetWidth < textNode?.scrollWidth;
-
-      const handleMouseEnter = (e: React.MouseEvent<HTMLButtonElement>) => {
-        const popoverParent = (e.target as HTMLButtonElement).closest(
-          `.${popoverClassName}`,
-        );
-        if (!popoverParent) setOpen(true);
-      };
-
-      const handleMouseLeave = () => setOpen(false);
-      const handleOnFocus = () => setOpen(true);
-      const handleOnBlur = () => setOpen(false);
-
-      return (
-        <div className={customMenuButtonWrapperStyles}>
-          <Tooltip
-            enabled={hasEllipse && !disabled}
-            justify="middle"
-            // Using refEl instead of a trigger element because triggerProps by default, such as onMouseEnter, are added to the trigger element inside the tooltip component. OnMouseEnter is triggered by hovering over the trigger or any of its children. In the case of this custom menu button we don't want the tooltip to open when children are hovered so we add our own open logic with onMouseEnter.
-            refEl={buttonRef}
-            open={open}
-            {...popoverProps}
-          >
-            {currentUnitOption?.displayName}
-          </Tooltip>
-          <Button
-            {...props}
-            className={cx(
-              menuBaseStyles,
-              menuThemeStyles[theme],
-              {
-                [menuThemeDisabledStyles[theme]]: disabled,
-              },
-              className,
-            )}
-            ref={buttonRef}
-            onMouseEnter={handleMouseEnter}
-            onMouseLeave={handleMouseLeave}
-            onFocus={handleOnFocus}
-            onBlur={handleOnBlur}
-          >
-            {children}
-          </Button>
-        </div>
-      );
-    },
+    ({ className, children, ...props }: ButtonProps, forwardedRef) => (
+      <SelectButton
+        ref={forwardedRef}
+        disabled={disabled}
+        displayName={currentUnitOption.displayName}
+        {...props}
+      >
+        {children}
+      </SelectButton>
+    ),
   );
-
   CustomMenuButton.displayName = 'CustomMenuButton';
 
   return (
@@ -144,7 +81,7 @@ export function Select({
         id={id}
         onChange={handleChange}
         aria-labelledby="Unit Picker"
-        value={currentUnitOption?.displayName as string}
+        value={currentUnitOption.displayName}
         className={cx(selectStyles, {
           [selectDisabledStyles[theme]]: disabled,
         })}
@@ -157,8 +94,8 @@ export function Select({
         __INTERNAL__menuButtonSlot__={CustomMenuButton}
       >
         {unitOptions?.map(option => (
-          <Option key={option?.displayName} value={option?.displayName}>
-            {option?.displayName}
+          <Option key={option.displayName} value={option.displayName}>
+            {option.displayName}
           </Option>
         ))}
       </SelectComponent>
