+++ packages/number-input/src/SelectButton/SelectButton.types.ts
@@ -1,0 +1,8 @@
+import { ButtonProps } from '@leafygreen-ui/button';
+
+import { PopoverProps } from '../Select/Select.types';
+
+export interface SelectButtonProps extends ButtonProps, PopoverProps {
+  disabled: boolean;
+  displayName: string;
+}
