+++ packages/number-input/src/SelectButton/index.tsx
@@ -1,0 +1,1 @@
+export { SelectButton } from './SelectButton';
